Download of full CSV of search results works now
It's faster because I'm not trying to upload to S3 and generate a
presigned URL, instead just letting the user download the CSV directly
from the server.

Also the table_search query is smarter and will return the exact number
of rows that you tell it to.

Still need to break all the DB connections in utils.py in the search app
out into their own class